refactor: 카테고리 application layer 패키지 이동 및 사용하지 않는 DTO 삭제
