--- backend/ddang/src/main/java/com/ddang/ddang/category/presentation/dto/response/ReadCategoriesResponse.java
@@ -1,6 +1,0 @@
-package com.ddang.ddang.category.presentation.dto.response;
-
-import java.util.List;
-
-public record ReadCategoriesResponse(List<ReadCategoryResponse> categories) {
-}
