+++ backend/ddang/src/main/java/com/ddang/ddang/category/presentation/dto/response/ReadCategoryResponse.java
@@ -1,6 +1,6 @@
 package com.ddang.ddang.category.presentation.dto.response;
 
-import com.ddang.ddang.category.application.dto.ReadCategoryDto;
+import com.ddang.ddang.category.application.dto.response.ReadCategoryDto;
 
 public record ReadCategoryResponse(Long id, String name) {
